requirements.txt
opencv-python==3.4.2.17
pyzmq==17.1.0
goprocam